+++ app/controllers/v1/groups_controller.rb
@@ -24,7 +24,7 @@ module V1
       @group = Group.create_with_image(
         group_params: group_params.merge(owner: current_user),
         invitees: params[:invitees],
-        group_image: group_image)
+        image: group_image)
 
       if @group.errors.empty?
         render :show, status: :created
@@ -40,8 +40,7 @@ module V1
       @group = Group.find(group_id)
       authorize @group
 
-      @group.update_with_image(group_params: group_params,
-                               group_image: group_image)
+      @group.update_with_image(group_params: group_params, image: group_image)
 
       if @group.errors.empty?
         render :show
@@ -182,12 +181,13 @@ module V1
 
     private def group_params
       p = params.permit(:name, :membersCanPost, :membersCanInvite)
-      p.transform_keys(&:underscore)
+      p.transform_keys(&:underscore) if p
     end
 
     private def group_image
       return nil unless params[:image].present?
-      params.require(:image).require(:publicId)
+      p = params.require(:image).permit(:publicId, :localId)
+      p.transform_keys(&:underscore.to_sym)
     end
 
     private def group_id